pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "movie-shot-categorizer"
version = "0.1.0"
authors = [
    { name="Hugging Face", email="team@huggingface.co" },
]
description = "A library to categorize movie shots using a Hugging Face model."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "transformers>=4.49.0",
    "torch",
    "Pillow",
    "accelerate>=1.0.0",
    "timm"
]

[project.urls]
"Homepage" = "https://github.com/huggingface/movie-shot-categorizer"
"Bug Tracker" = "https://github.com/huggingface/movie-shot-categorizer/issues"